Reactor stale files in components
