--- src/components/Buttons/index.css
@@ -1,133 +1,0 @@
-:root {
-  --ct-btn-green: #328383;
-  --ct-btn-icon-green: #328383;
-}
-
-.ct-btn {
-  background: transparent;
-  border: 0;
-  padding: 0;
-  margin: 0;
-  outline: none;
-  text-align: left;
-  color: inherit;
-  position: relative;
-}
-.ct-btn .ct-btn-content {
-  display: flex;
-  flex-direction: row;
-  border-radius: .28571429rem;
-  overflow: hidden;
-  /* transition: filter .2s ease-in; */
-}
-.ct-btn:hover > .ct-btn-content {
-  filter: brightness(110%);
-}
-.ct-btn:focus > .ct-btn-content {
-  filter: brightness(110%);
-}
-.ct-btn:active > .ct-btn-content {
-  filter: brightness(90%);
-}
-.ct-btn:disabled > .ct-btn-content {
-  opacity: .7;
-  pointer-events: none;
-  filter: none;
-}
-.ct-btn .ct-btn-text {
-  color: inherit;
-  background: inherit;
-  font-family: var(--ct-font-google);
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: center;
-}
-
-.ct-btn .ct-btn-icon {
-  display: block;
-  background: #0000001f;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: center;
-  padding: 0 .5em;
-  height: 2.5em;
-}
-
-.ct-btn[data-circle=true] .ct-btn-content {
-  width: 35px;
-  height: 35px;
-  border-radius: 50%;
-  padding: 0;
-  overflow: auto;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-.ct-btn[data-circle=true] .ct-btn-icon {
-  padding: 0;
-  width: 35px;
-  height: 35px;
-  border-radius: 50%;
-}
-
-/* Color */
-.ct-btn[data-color="green"] .ct-btn-content {
-  background: var(--ct-btn-green);
-  color: white;
-}
-.ct-btn[data-color="primary"] .ct-btn-content {
-  background: rgba(209, 209, 209, 0.521);
-  color: rgb(66, 66, 66);
-}
-.ct-btn[data-color="light"] .ct-btn-content {
-  background: rgb(253, 253, 253);
-  color: rgb(77, 77, 77);
-}
-.ct-btn[data-color="yellow"] .ct-btn-content {
-  background: rgb(246, 201, 0);
-  color: black;
-}
-.ct-btn[data-color="red"] .ct-btn-content {
-  background: #db2828;
-  color: white;
-}
-.ct-btn[data-color="text-green"] .ct-btn-content {
-  color: rgb(0, 128, 117);
-}
-.ct-btn[data-color="text-green"] .ct-btn-content:hover {
-  color: rgb(5, 131, 120);
-}
-
-
-/* Size */
-.ct-btn[data-font*=normal] .ct-btn-text {
-  font-size: 1.2em;
-  padding: .5em 1.2em;
-}
-.ct-btn[data-font*=big] .ct-btn-text {
-  font-size: 1.4em;
-  padding: .5em 1.2em;
-}
-.ct-btn[data-font*=big] .ct-btn-icon {
-  height: 3em;
-}
-.ct-btn[data-font*=bold] .ct-btn-text {
-  font-weight: bold;
-}
-
-
-
-
-/* Group */
-.ct-btn-group .ct-btn {
-  margin: 0 .3em;
-}
-.ct-btn-group .ct-btn:first-child {
-  margin-left: 0;
-}
-.ct-btn-group .ct-btn:last-child {
-  margin-right: 0;
-}
