--- src/components/Buttons/index.js
@@ -1,77 +1,0 @@
-import React from 'react';
-import './index.css';
-import { Popup } from 'semantic-ui-react';
-
-export function CTButton({
-  id = `ct-btn-${Math.random()}`,
-  type = 'button',
-  classNames = '',
-  spanClassNames = '',
-  color = 'primary', // primary, light, green, yellow, black
-  size = 'normal',
-  icon = null,
-  iconContent = null,
-  circle = false,
-
-  onClick = null,
-  text = '',
-  disabled = false,
-
-  popup = null,
-  popupAttrs = {},
-  popupBasic = true,
-  popupDelay = 0,
-
-  title,
-  ariaLabel,
-  ariaHasPopup,
-  ariaControls,
-
-  children,
-  ...others
-}) {
-  return (
-    <Popup
-      inverted
-      mouseEnterDelay={popupDelay}
-      basic={popupBasic}
-      disabled={!popup}
-      openOnTriggerFocus
-      closeOnTriggerBlur
-      {...popupAttrs}
-      content={popup}
-      trigger={
-        <button
-          id={id}
-          type={type}
-          className={`ct-btn ${classNames}`}
-          data-font={size}
-          data-color={color}
-          data-icon={Boolean(icon).toString()}
-          data-circle={circle.toString()}
-          disabled={disabled}
-          onClick={onClick}
-          title={title}
-          aira-label={ariaLabel || text}
-          aria-haspopup={ariaHasPopup}
-          aria-controls={ariaControls}
-          {...others}
-        >
-          <span className="ct-btn-content" tabIndex="-1">
-            {Boolean(icon) && (
-              <span className="ct-btn-icon" aria-hidden="true">
-                {iconContent || <i className="material-icons">{icon}</i>}
-              </span>
-            )}
-            {Boolean(text) && (
-              <span className={`ct-btn-text ${spanClassNames}`} aria-hidden="true">
-                {text}
-                {children}
-              </span>
-            )}
-          </span>
-        </button>
-      }
-    />
-  );
-}
