+++ src/components/Modals/index.scss
@@ -2,45 +2,6 @@
  * Stylesheet for General Modal
  */
 
-.ui .modal {
-  overflow: visible;
-  box-shadow: none !important;
-}
-.ui.modal>.header, .ui.modal>.actions {
-  background: #fcfcfc !important;
-  border: none !important;
-}
-.solid-dimmer {
-  transition: none !important;
-  background-color: #fcfcfc !important;
-}
-
-.general-modal {
-  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
-  position: relative !important;
-  height: max-content;
-  display: flex;
-  flex-direction: column;}
-  .general-modal .gm-header {
-    font-family: Impact, Haettenschweiler, 'Arial Narrow Bold', sans-serif !important;
-    background: #eeeeee !important; 
-    color: #217272 !important;}
-  .general-modal .gm-content {
-    background: #fcfcfc !important;}
-  .general-modal .gm-action {
-    background: #f0f0f0 !important;}
-
-  /* For Delete Modalc*/   
-  .general-modal .delete-target {
-    color: rgb(26, 110, 110);
-    text-decoration: underline;
-  }
-
-  .general-modal .positive {
-    background: #449999 !important;
-  }
-
-
 .ct-modal {
   position: fixed;
   top: 0;
