+++ src/screens/Admin/Alerts/index.css
