+++ src/screens/Admin/index.css
@@ -136,3 +136,42 @@
   color: white;
   border: none !important;
 }
+
+
+.ui .modal {
+  overflow: visible;
+  box-shadow: none !important;
+}
+.ui.modal>.header, .ui.modal>.actions {
+  background: #fcfcfc !important;
+  border: none !important;
+}
+.solid-dimmer {
+  transition: none !important;
+  background-color: #fcfcfc !important;
+}
+
+.general-modal {
+  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
+  position: relative !important;
+  height: max-content;
+  display: flex;
+  flex-direction: column;}
+  .general-modal .gm-header {
+    font-family: Impact, Haettenschweiler, 'Arial Narrow Bold', sans-serif !important;
+    background: #eeeeee !important; 
+    color: #217272 !important;}
+  .general-modal .gm-content {
+    background: #fcfcfc !important;}
+  .general-modal .gm-action {
+    background: #f0f0f0 !important;}
+
+  /* For Delete Modalc*/   
+  .general-modal .delete-target {
+    color: rgb(26, 110, 110);
+    text-decoration: underline;
+  }
+
+  .general-modal .positive {
+    background: #449999 !important;
+  }
